Refactor: Remove duplication
